+++ dpws/src/main/java/org/somda/sdc/dpws/DpwsConstants.java
@@ -167,8 +167,8 @@ public class DpwsConstants {
             "(\\.(25[0-5]|2[0-4]\\d|1\\d\\d|[1-9]?\\d)){3}))|:)))" +
             "(%[a-fA-F0-9]{2})?\\s*";
     private static final String IP_LITERAL = "\\[" + "((" + IPV6_ADDRESS + ")|(" + IPV_FUTURE + "))" + "\\]";
-    private static final String HOST = "({host}(" + REG_NAME + "|" + IPV4_ADDRESS + "|" + IP_LITERAL + "))";
-    public static final String AUTHORITY = "(({userInfo}" + USER_INFO + ")@)?" + HOST + "(:({port}[0-9]*))?";
+    private static final String HOST = "({=host}(" + REG_NAME + "|" + IPV4_ADDRESS + "|" + IP_LITERAL + "))";
+    public static final String AUTHORITY = "(({=userInfo}" + USER_INFO + ")@)?" + HOST + "(:({=port}[0-9]*))?";
     private static final String PATH_EMPTY = "";
     private static final String PATH_ROOTLESS = SEGMENT_NZ_REGEX + "(/" + SEGMENT_REGEX + ")*";
     private static final String PATH_NOSCHEME = "[a-zA-Z0-9\\-._~!$&'()*+,;=@]+" + "(/" + SEGMENT_REGEX + ")*";
@@ -184,24 +184,36 @@ public class DpwsConstants {
     private static final String QUERY = "(" + P_CHAR + "|/|\\?)*";
     private static final String FRAGMENT = QUERY;
 
+    private static final String PARAM = P_CHAR + "*";
+    private static final String SEGMENT = P_CHAR + "*(;" + PARAM + ")*";
+    private static final String PATH_SEGMENTS = SEGMENT + "(/" + SEGMENT + ")*";
+
+    private static final String ESCAPED = "%" + HEXDIG + HEXDIG;
+    private static final String RESERVED = "[;/?:@&=+$,]";
+    private static final String UNRESERVED = "[a-zA-Z0-9\\-_.!~*'()]";
+    private static final String REL_SEGMENT = "(" + UNRESERVED + "|" + ESCAPED + "|[;@&=+$,])+";
+
+    private static final String ABS_PATH = "({=path}/" + PATH_SEGMENTS + ")";
+    private static final String NET_PATH = "//({=authority}" + AUTHORITY + ")(" + ABS_PATH + ")?";
+    private static final String REL_PATH = REL_SEGMENT + "(" + ABS_PATH + ")?";
+
+    public static final String RELATIVE_URI = "(" + NET_PATH + "|" + ABS_PATH + "|" + REL_PATH + ")(\\?({=query}" + QUERY + "))?";
+
+    private static final String URIC = RESERVED + "|" + UNRESERVED + "|" + ESCAPED;
+    private static final String URIC_NO_SLASH = UNRESERVED + "|" + ESCAPED + "|" + "[;?:@&=+$,]";
+    private static final String OPAQUE_PART = URIC_NO_SLASH + URIC + "*";
+    private static final String HIER_PART = "(" + NET_PATH + "|" + ABS_PATH + ")(\\?({=query}" + QUERY + "))?";
+
+    public static final String ABSOLUTE_URI = "({scheme}" + SCHEME_SEGMENT + ")" +
+            ":" + "(" + HIER_PART + "|" + OPAQUE_PART + ")";
+
+    public static final String URI_REFERENCE = "(({absoluteUri}" + ABSOLUTE_URI + ")|({relativeUri}" + RELATIVE_URI + "))(#({fragment}" + FRAGMENT + "))?";
+
     // Added negative lookahead for "//" to prevent authority from being interpreted as path
     public static final String URI_REGEX = "^(" +
             "({scheme}" + SCHEME_SEGMENT + ")" +
             ":" +
             "(//({authority}" + AUTHORITY + "))?" +
-            "({path}" +
-            "(?(authority)" +
-            "(((?=/)" + PATH + ")|)|((?!//)" + PATH + ")" +
-            ")" +
-            ")" +
-            "(\\?({query}" + QUERY + "))?" +
-            "(#({fragment}" + FRAGMENT + "))?" +
-            ")$";
-
-    public static final String RELATIVE_URI_REGEX = "^(" +
-            "(({scheme}" + SCHEME_SEGMENT + ")" +
-            ":" +
-            "//)?({authority}" + AUTHORITY + ")?" +
             "({path}" +
             "(?(authority)" +
             "(((?=/)" + PATH + ")|)|((?!//)" + PATH + ")" +